moved BufferInterface to Contracts
